cambio tipo de dato dni
se cambia el dni de long a long long y se corrige el bug de los 11 digitos.

+++ VentasManager.h
@@ -25,10 +25,10 @@ public:
 	void restaurarBackup();
 	void borrarVenta();
 	
-	int validarCliente(long dni);
+	int validarCliente(long long dni);
 	void crearNuevoCliente();
 	void mostrarClienteAsociado(int pos);
-	std::string mostrarNombreCliente(long dni);
+	std::string mostrarNombreCliente(long long dni);
 	
 	int validarSucursal(int id);
 	void mostrarSucursalAsociada(int pos);
